backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
qdrant-client==1.12.1
openai==1.58.1
python-dotenv==1.0.1
watchdog==6.0.0
langchain-text-splitters==0.3.3
pydantic==2.10.3
websockets==14.1
